requirements_backup.txt
supervisely==6.72.55
pre-commit
yapf==0.26.0
flake8
pyyaml>=5.1
visualdl>=2.0.0
opencv-python
tqdm
filelock
scipy
prettytable
eiseg
paddlepaddle
pyqt5
qtpy
paddleseg
albumentations
cython
wget
diskcache==5.2.1